macos: move grab handle to separate file
